mishwari_server/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
stripe==9.10.0
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
googlemaps==4.10.0
geopy==2.4.1
numpy==1.26.4
Pillow==10.1.0
twilio==9.2.3
polyline==2.0.2
shapely==2.0.3